Create README.md
This repository contains the problem of gender identification and the solution of that problem provided with the datasets. I have applied different Machine Learning algorithms on it and classify the record as male or female. Pick up the best model and save as pickle file to classify the records in future.